feat: 추천 섹션 로직을 커스텀 훅으로 분리 및 컴포넌트 모듈화
